bygg: remove buggy skip-init-build
When do we need to skip the things in init_yocto_build? where is the code that leave us in the build folder?

If i use 	export BUILD_COMMAND=''bash" withe the build i will not get the init_yocto_build if  i exit and restart the bygg without that export.