bygg: remove buggy skip-init-build
When do we need to skip the things in init_yocto_build? where is the code that leave us in the build folder? I did not get this to work at all.

If i use export BUILD_COMMAND=''bash" with the build I will not get the init_yocto_build sometimes and if I exit this shell and restart the bygg without that export it will not add the init_yocto_build.

using BUILD_COMMAND whit bash do the same thing leave us in build dir with all ENV that is included in the script and from outside.